+++ lib/main.dart
@@ -21,7 +21,6 @@ import 'package:MobileOne/localization/supported.dart';
 import 'package:MobileOne/pages/Mainpage.dart';
 import 'package:MobileOne/pages/create_list.dart';
 import 'package:MobileOne/pages/forgotten_password.dart';
-import 'package:MobileOne/pages/lists.dart';
 import 'package:MobileOne/pages/openedListPage.dart';
 import 'package:MobileOne/pages/register-page.dart';
 import 'package:MobileOne/services/analytics_service.dart';
@@ -100,7 +99,6 @@ class MyApp extends StatelessWidget {
         '/': (context) => AuthenticationPage(),
         '/registerPage': (context) => RegisterPage(),
         '/forgottenPasswordPage': (context) => ForgottenPasswordPage(),
-        '/lists': (context) => Lists(),
         '/createList': (context) => CreateList(),
         '/openedListPage': (context) => OpenedListPage(),
         '/shareOne': (context) => ShareOne(),